test: move test utils into helpers folder
